refactor(frontend): extract shared escapeHtml utility from duplicated implementations (#3215)
